Add test_queue to tests
